Move pom.xml to repo root, simplify structure
Running a local becomes easier when the main pom file resides in the
root directory of the git repo.

Move the pom.xml to the repository root. Get rid of the chart.build
project that only contained the pom.xml and a nested maven module.
Make this nested repository module a project of its own.

Change-Id: I7df311230a6f34f165b3807c46ca2ff826a55300