Add Stacked/Grouped Multi-Bar Chart widget and snippet
Included new kind of widget named Stacked/Grouped Multi-Bar Chart that
contains multiple bar groups with the different values and options in
the same
graph. There is two kind options, one is for widget options and another
is
for bar group option. 

Create a basic StackedMultiBarChart based on nv's multiBarChart and
include nv-stacked-multi-bar JavaScript file.
Create StackedMultiBarChart_Test.
Add a snippet, and include it in ChartSnippetPage.

Change-Id: Icc9388570f9dfb8b29d1717fdef2095de6e5aff8
Signed-off-by: Elshad Seyidmammadov <elshad@eclipsesource.com>