Add HistoricalBarChart widget and snippet
Include new widget named HistoricalBarChart that contains single bar
type with the different values. The chart is like to LineChart but the
graph is described only with the bars and also StackedMultiBarChart with
the single type of bar is look like the same with it.


Create a basic HistoricalBarChart based on nv's historicalBarChart and
include nv-historical-bar JavaScript file.
Create HistoricalBarChart_Test.
Add a snippet, and include it in ChartSnippetPage.


Change-Id: Ifff58e93194feb3cf6b6f032d56c51ec4de0d4fe
Signed-off-by: Elshad Seyidmammadov <elshad@eclipsesource.com>